casted center as int

+++ Spaceship.pde
@@ -1,9 +1,9 @@
 class Spaceship extends Floater  
 {   
   public void setX(int x){myCenterX = x;} 
-  public int getX(){return myCenterX;}   
+  public int getX(){return (int) myCenterX;}   
   public void setY(int y){myCenterY = y;}   
-  public int getY(){return myCenterY;}   
+  public int getY(){return (int) myCenterY;}   
   public void setDirectionX(double x){myDirectionX = x;}   
   public double getDirectionX() {return myDirectionX;}
   public void setDirectionY(double y){myDirectionY = y;}
